Remove NetworkManager from Vagrant box
NetworkManager is problematic in Vagrant environments since it will
tear down hostonly private networks as soon as they are created by
the vagrant provisioner. This patch removes it from both image builds
and from the dnsmasq role. Additionally it adds a route required
by ansible to determine the default ipv4 address for a node.